refactor(server): isolate discovery implementation
